实现接口信息删除功能、优化UI

+++ openapi-platform-frontend/src/pages/InterfaceInfo/index.tsx
@@ -1,34 +1,33 @@
 import {
   addInterfaceInfoUsingPost,
+  deleteInterfaceInfosUsingPost,
+  deleteInterfaceInfoUsingPost,
   listInterfaceInfoByPageUsingPost,
   updateInterfaceInfoUsingPost,
 } from '@/services/backend-server/interfaceInfoController';
-import { PlusOutlined } from '@ant-design/icons';
+import { ExclamationCircleFilled, PlusOutlined } from '@ant-design/icons';
 import type { ActionType, ProColumns } from '@ant-design/pro-components';
 import { FooterToolbar, PageContainer, ProTable } from '@ant-design/pro-components';
-import { Button, message } from 'antd';
+import { Button, message, Modal } from 'antd';
 import React, { useRef, useState } from 'react';
 import CreateModal from './components/CreateModal';
 import UpdateModal from './components/UpdateModal';
+import './index.less';
+
+const { confirm } = Modal;
 
 const TableList: React.FC = () => {
-  /**
-   * @en-US Pop-up window of new window
-   * @zh-CN 新建窗口的弹窗
-   *  */
   const [createModalOpen, handleModalOpen] = useState<boolean>(false);
-  /**
-   * @en-US The pop-up window of the distribution update window
-   * @zh-CN 分布更新窗口的弹窗
-   * */
+
   const [updateModalOpen, handleUpdateModalOpen] = useState<boolean>(false);
 
   const [showDetail, setShowDetail] = useState<boolean>(false);
 
   const actionRef = useRef<ActionType>();
   const [currentRow, setCurrentRow] = useState<API.InterfaceInfo>();
-  const [selectedRowsState, setSelectedRows] = useState<API.RuleListItem[]>([]);
+  const [selectedRowsState, setSelectedRows] = useState<API.InterfaceInfo[]>([]);
 
+  // 添加
   const handleAdd = async (fields: API.InterfaceInfo) => {
     const hide = message.loading('正在添加');
     try {
@@ -44,10 +43,10 @@ const TableList: React.FC = () => {
     }
   };
 
+  // 修改
   const handleUpdate = async (fields: API.InterfaceInfo) => {
     const hide = message.loading('修改中');
     try {
-      console.log(fields);
       await updateInterfaceInfoUsingPost({
         ...fields,
       });
@@ -62,27 +61,67 @@ const TableList: React.FC = () => {
     }
   };
 
+  // 删除多个数据
+  const handleRemove = async (items: API.InterfaceInfo[]) => {
+    let ids = items.map((item) => item.id as number).filter((id) => id !== undefined);
+    let delReq: API.InterfaceDeleteRequest = {
+      ids: ids,
+    };
+    const hide = message.loading('删除中');
+    try {
+      hide();
+      await deleteInterfaceInfosUsingPost(delReq);
+      message.success('删除成功');
+      return true;
+    } catch (err: any) {
+      hide();
+      message.error('删除异常' + err.message);
+      return false;
+    }
+  };
+
+  // 删除单个数据
+  const handleRemoveOne = async (item: API.InterfaceInfo) => {
+    const hide = message.loading('删除中');
+    try {
+      hide();
+      await deleteInterfaceInfoUsingPost({ id: item.id });
+      message.success('删除成功');
+      return true;
+    } catch (err: any) {
+      hide();
+      message.error('删除失败' + err.message);
+      return false;
+    }
+  };
+
   const columns: ProColumns<API.InterfaceInfo>[] = [
     {
       title: 'id',
       dataIndex: 'id',
       valueType: 'index',
+      // fixed: 'left',
+      width: 80,
     },
     {
       title: '接口名称',
       dataIndex: 'name',
       valueType: 'text',
+      // fixed: 'left',
+      width: 160,
     },
     {
       title: '描述',
       dataIndex: 'description',
       valueType: 'textarea',
       ellipsis: true,
+      width: 300,
     },
     {
       title: '请求方法',
       dataIndex: 'method',
       valueType: 'text',
+      width: 80,
     },
     {
       title: '请求地址',
@@ -90,6 +129,7 @@ const TableList: React.FC = () => {
       valueType: 'text',
       copyable: true,
       ellipsis: true,
+      width: 260,
     },
     {
       title: '请求头',
@@ -97,6 +137,8 @@ const TableList: React.FC = () => {
       valueType: 'textarea',
       ellipsis: true,
       copyable: true,
+      width: 360,
+      hideInSearch: true,
     },
     {
       title: '响应头',
@@ -104,11 +146,14 @@ const TableList: React.FC = () => {
       valueType: 'textarea',
       ellipsis: true,
       copyable: true,
+      width: 360,
+      hideInSearch: true,
     },
     {
       title: '接口状态',
       dataIndex: 'status',
       hideInForm: true,
+
       valueEnum: {
         0: {
           text: '关闭',
@@ -119,6 +164,7 @@ const TableList: React.FC = () => {
           status: 'Processing',
         },
       },
+      width: 80,
     },
     {
       title: '创建时间',
@@ -126,6 +172,8 @@ const TableList: React.FC = () => {
       valueType: 'dateTime',
       hideInForm: true,
       ellipsis: true,
+      width: 160,
+      hideInSearch: true,
     },
     {
       title: '更新时间',
@@ -133,37 +181,75 @@ const TableList: React.FC = () => {
       valueType: 'dateTime',
       hideInForm: true,
       ellipsis: true,
+      width: 160,
+      hideInSearch: true,
     },
     {
       title: '操作',
       dataIndex: 'option',
       valueType: 'option',
+      width: 160,
       render: (_, record) => [
         <a
-          key="config"
+          key="edit"
           onClick={() => {
             handleUpdateModalOpen(true);
             setCurrentRow(record);
           }}
+          className="warning-text"
         >
           修改
         </a>,
-        <a key="subscribeAlert" href="https://procomponents.ant.design/">
+        <a key="detail" href="https://procomponents.ant.design/">
           查看
+        </a>,
+        <a
+          key="delete"
+          className="danger-text"
+          onClick={() => {
+            confirm({
+              title: '操作警告',
+              icon: <ExclamationCircleFilled />,
+              content: '数据删除后不可逆，确定要删除选中的这条数据吗？',
+              okText: '确定',
+              okType: 'danger',
+              cancelText: '取消',
+              onOk: async () => {
+                const success = await handleRemoveOne(record);
+                if (success) {
+                  if (actionRef.current) {
+                    actionRef.current.reload();
+                  }
+                }
+              },
+              onCancel() {
+                console.log('Cancel');
+              },
+            });
+          }}
+        >
+          删除
         </a>,
       ],
     },
   ];
 
+  type PageParams = {
+    current: number;
+    pageSize: number;
+  };
   return (
     <PageContainer>
-      <ProTable<API.RuleListItem, API.PageParams>
+      <ProTable<API.InterfaceInfo, PageParams>
+        scroll={{
+          x: 2400,
+        }}
         cardBordered={{
           table: true,
         }}
         headerTitle="接口详情"
         actionRef={actionRef}
-        rowKey="key"
+        rowKey="id"
         search={{
           labelWidth: 120,
         }}
@@ -219,11 +305,25 @@ const TableList: React.FC = () => {
           }
         >
           <Button
-            onClick={async () => {
-              await handleRemove(selectedRowsState);
-              setSelectedRows([]);
-              actionRef.current?.reloadAndRest?.();
+            onClick={() => {
+              confirm({
+                title: '操作警告',
+                icon: <ExclamationCircleFilled />,
+                content: '数据删除后不可逆，确定要删除选中的这些数据吗？',
+                okText: '确定',
+                okType: 'danger',
+                cancelText: '取消',
+                onOk: async () => {
+                  await handleRemove(selectedRowsState);
+                  setSelectedRows([]);
+                  actionRef.current?.reloadAndRest?.();
+                },
+                onCancel() {
+                  console.log('Cancel');
+                },
+              });
             }}
+            className="danger-bg"
           >
             批量删除
           </Button>
